+++ src/RedAnt.java
@@ -1,7 +1,7 @@
 public class RedAnt extends Ant {
     private int capacity;
 
-    public RedAnt(int id, int x, int y, int capacity) {
+    public RedAnt(int id, float x, float y, int capacity) {
         super(id, x, y);
         this.capacity = capacity; // red ant capacity
     }
